python/requirements.txt
flask>=3.0
python-dotenv>=1.0
requests>=2.31
pandas>=2.0
openpyxl>=3.1